end of sec 14

+++ Domain/Activity.cs
@@ -3,15 +3,15 @@
     public class Activity
     {
         public string Id { get; set; } = Guid.NewGuid().ToString();
-        public string Title { get; set; }=string.Empty;
+        public string Title { get; set; } = string.Empty;
         public string Category { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public bool IsCancelled { get; set; }
         public DateTime Date { get; set; }
         public string City { get; set; } = string.Empty;
         public string Venue { get; set; } = string.Empty;
-        public int longitude { get; set; }
-        public int latitude { get; set; }
+        public double longitude { get; set; }
+        public double latitude { get; set; }
         public ICollection<ActivityAttendee> attendees { get; set; } = [];
     }
 }
